Refactor whiteboard functionality into GameComponent and share WebSocket
